pyproject.toml
[tool.poetry]
name = "repl_nix_adventofcode"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
requests = "^2.28.1"
rich = "^12.6.0"
docutils = "^0.19"
python-dotenv = "^0.21.0"
django-render = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
